create-itp-react-app: split up index into setupGenerator & runGenerator
