+++ packages/create-itp-react-app/lib/logTitle.js
@@ -1,0 +1,6 @@
+const chalk = require('chalk');
+
+module.exports = title => {
+  console.log(' ');
+  console.log(chalk.inverse(title));
+};
